\name{epwpl-package}
\alias{epwpl-package}
\alias{epwpl}
\docType{package}
\title{
\packageTitle{epwpl}
}
\description{
\packageDescription{epwpl}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{epwpl}
\packageIndices{epwpl}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{epwpl}

Maintainer: \packageMaintainer{epwpl}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
